requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.36
alembic==1.13.2
PyYAML==6.0.2
prometheus_client==0.21.0
httpx==0.27.2
websockets==12.0
pytest==8.3.3
pytest-asyncio==0.24.0
coverage==7.6.4
aiosqlite==0.20.0
black==24.8.0
mypy==1.10.0
ruff==0.6.5
